processor/src/main/java/io/bothy/tog/processor/BuilderField.java
/*
 * Copyright © 2024 Neil Richard Green
 *
 * This file is part of Tog.
 *
 * Tog is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as
 * published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
 *
 * Tog is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License along with Foobar. If not, see
 * <https://www.gnu.org/licenses/>.
 */
package io.bothy.tog.processor;

import javax.lang.model.element.Name;
import javax.lang.model.element.RecordComponentElement;
import javax.lang.model.element.VariableElement;
import javax.lang.model.type.TypeMirror;

record BuilderField(Name fieldName, TypeMirror fieldType) {
    static BuilderField from(final RecordComponentElement recordComponent) {
        return new BuilderField(recordComponent.getSimpleName(), recordComponent.asType());
    }

    public static BuilderField from(final VariableElement variable) {
        return new BuilderField(variable.getSimpleName(), variable.asType());
    }
}
